same stuff

+++ react-next/08-crd/src/app/alunos/page.js
@@ -5,24 +5,42 @@ import { useEffect, useState } from 'react';
 import { Button, Table } from 'react-bootstrap';
 import { FaPenAlt, FaPlusCircle } from "react-icons/fa";
 import { FaX } from 'react-icons/fa6';
-import { toast } from 'react-toastify';
 
 export default function AlunosPage() {
 
   const [alunos, setAlunos] = useState([])
+  const [faculdades, setFaculdades] = useState([])
+  const [cursos, setCursos] = useState([])
 
   useEffect(() => {
     const alunosLocalStorage = JSON.parse(localStorage.getItem('alunos')) || [];
-    console.log("🚀 ~ useEffect ~ alunosLocalStorage:", alunosLocalStorage)
     setAlunos(alunosLocalStorage);
+
+    const faculdadesLocalStorage = JSON.parse(localStorage.getItem('faculdades')) || [];
+    setFaculdades(faculdadesLocalStorage);
+
+    const cursosLocalStorage = JSON.parse(localStorage.getItem('cursos')) || [];
+    setCursos(cursosLocalStorage);
+
   }, []);
 
-
   function excluir(aluno) {
-    const novosAlunos = alunos.filter((item) => item.id !== aluno.id);
-    localStorage.setItem('alunos', JSON.stringify(novosAlunos));
-    setAlunos(novosAlunos);
-    toast.success(`Aluno ${aluno.nome} excluído com sucesso!`);
+    if (window.confirm(`Deseja realmente excluir o curso ${aluno.nome}?`)) {
+      const novosAlunos = alunos.filter((item) => item.id !== aluno.id);
+      localStorage.setItem('alunos', JSON.stringify(novosAlunos));
+      setCursos(novosAlunos);
+      alert(`Aluno ${aluno.nome} excluído com sucesso!`);
+    }
+  }
+
+  function getCurso(id) {
+    const curso = cursos.find((item) => item.id === id);
+    return curso ? curso.nome : 'Desconhecido';
+  }
+
+  function getFaculdade(id) {
+    const faculdade = faculdades.find((item) => item.id === id);
+    return faculdade ? faculdade.nome : 'Desconhecido';
   }
 
   return (
@@ -50,11 +68,11 @@ export default function AlunosPage() {
               <td>{aluno.matricula}</td>
               <td>{aluno.nome}</td>
               <td>{aluno.sobrenome}</td>
-              <td>{aluno.curso}</td>
-              <td>{aluno.faculdade}</td>
+              <td>{getCurso(aluno.curso)}</td>
+              <td>{getFaculdade(aluno.faculdade)}</td>
               <td>{aluno.periodo}</td>
               <td className='text-center'>
-                <Button variant="warning" href={`/alunos/form/${aluno.id}`} title="Editar"><FaPenAlt /></Button>
+                <Button variant="warning" href={`/alunos/form?id=${aluno.id}`} title="Editar"><FaPenAlt /></Button>
                 <Button variant="danger" className="ms-2" title="Excluir" onClick={() => excluir(aluno)}><FaX /></Button>
               </td>
             </tr>
